Extract quote append extraction lifecycle slice
