+++ backend/app/features/quotes/extraction_append/__init__.py
@@ -1,0 +1,5 @@
+"""Quote extraction append lifecycle slice."""
+
+from app.features.quotes.extraction_append.service import QuoteExtractionAppendService
+
+__all__ = ["QuoteExtractionAppendService"]
